requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
drf-yasg==1.16.1
Flask==1.0.3
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.3
python-memcached==1.59
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.100
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
